requirements.txt
django-5.1.2